Time: 13 ms (74.17%) | Memory: 9.1 MB (85.58%) - LeetSync

+++ 383-ransom-note/ransom-note.cpp
@@ -3,7 +3,7 @@ public:
     bool canConstruct(string ransomNote, string magazine) {
         if (ransomNote.size() > magazine.size()) {return (false);}
         int i;
-        int16_t lowercases[26] = {0};
+        int lowercases[26] = {0};
         for (i = 0; magazine[i] != '\0'; i++)
         {
             lowercases[magazine[i] - 'a']++;
